Quick edit to fix the Sigwada's Star Card
Custom characters that weren't using Carol's event activator were caught in a weird state where the Star Card she uses for it would appear, but so would the character base's cutscene trigger.

If the player was fast enough, they could collect the card before the cutscene, leading to a functional Results Screen overlapping the Cory fight.

A more sensible way to do this might be to add a thing to the playable character class to specify scenes to remove things from? Although some Star Cards would need some hardcoding shenanigans (like with the Sigwada, which has two).