Update PatchBakunawaFusion.cs
The edit to add everyone to the assist roster was mistakenly checking for LESS than or equal to 5, so playing as a base game character would have themselves show up in their part to assist. While playing as a modded character beyond 5 would (in theory) have NO ONE show up.